+++ src/dgscript/dg_event.c
@@ -22,7 +22,7 @@
 #include "structs.h"
 #include "utils.h"
 #include "db.h"
-#include "dg_event.h"
+#include "dg_event_internal.h"
 #include "constants.h"
 #include "comm.h" /* For access to the game pulse */
 #include "dotenv.h"